+++ s2gos-client/src/tests/cli/test_config.py
@@ -1,6 +1,7 @@
 #  Copyright (c) 2025 by ESA DTE-S2GOS team and contributors
 #  Permissions are hereby granted under the terms of the Apache 2.0 License:
 #  https://opensource.org/license/apache-2-0.
+
 import os
 import unittest
 from pathlib import Path
@@ -12,7 +13,7 @@ import pytest
 from s2gos_client import ClientConfig
 from s2gos_client.api.defaults import DEFAULT_CONFIG_PATH
 from s2gos_client.cli.config import configure_client, get_config
-from s2gos_common.testing import set_env
+from s2gos_common.testing import set_env, set_env_cm
 
 DEFAULT_CONFIG_BACKUP_PATH = DEFAULT_CONFIG_PATH.parent / (
     str(DEFAULT_CONFIG_PATH.name) + ".backup"
@@ -98,3 +99,30 @@ class ReadConfigTest(unittest.TestCase):
         finally:
             if custom_config_path.exists():
                 os.remove(custom_config_path)
+
+    @patch("typer.prompt")
+    def test_configure_client_use_defaults(self, mock_prompt):
+        # Simulate sequential responses to typer.prompt
+        with set_env_cm(
+            S2GOS_USER_NAME="bibo",
+            S2GOS_ACCESS_TOKEN="9823hc",
+            S2GOS_SERVER_URL="http://localhorst:2357",
+        ):
+            mock_prompt.side_effect = [None, None, None]
+            custom_config_path = Path("test.cfg")
+            try:
+                actual_config_path = configure_client(config_path=custom_config_path)
+                self.assertEqual(custom_config_path, actual_config_path)
+                self.assertTrue(custom_config_path.exists())
+                config = get_config(custom_config_path)
+                self.assertEqual(
+                    ClientConfig(
+                        user_name="bibo",
+                        access_token="9823hc",
+                        server_url="http://localhorst:2357",
+                    ),
+                    config,
+                )
+            finally:
+                if custom_config_path.exists():
+                    os.remove(custom_config_path)